+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/items/altimeter/AltimeterItem.java
@@ -17,9 +17,9 @@ public class AltimeterItem extends Item {
     @Override
     public InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand hand) {
         ItemStack stack = player.getItemInHand(hand);
-        for (int i = level.getMinBuildHeight(); i <= level.getMaxBuildHeight(); i += 32) {
+        for (int i = level.getMinBuildHeight(); i <= level.getMaxBuildHeight(); i += 16) {
             if (player.getBlockY() < i) {
-                player.displayClientMessage(Component.literal("%d ~ %d".formatted(i - 32, i)).withStyle(ChatFormatting.GRAY), true);
+                player.displayClientMessage(Component.literal("%d ~ %d".formatted(i - 16, i)).withStyle(ChatFormatting.GRAY), true);
                 break;
             }
         }